Add ability to edit command before recompile and a function to rerun
See #77

This change does three things:

First it adds a prefix argument to the `recompile` and `rerun` commands so the cargo command can be
edited before running.

Second it adds a new function `rustic-cargo-test-rerun-current`, that can be called from a
`rustic-cargo-test-mode` buffer to rerun the test at point. Bound to `C-c C-t` by default.

Third, it adds minibuffer history support to the cargo-run-* and cargo-test-* commands.